AntiMatter/requirements.txt
torch==2.0.1
transformers==4.30.2
tokenizers==0.13.3
datasets==2.12.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0
wandb==0.15.4
tensorboard==2.13.0
pyyaml==6.0
ftfy==6.1.1
sentencepiece==0.1.99
sacremoses==0.0.53
nltk==3.8.1
spacy==3.5.3
regex==2023.6.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.2
jsonlines==3.1.0
accelerate==0.20.3
bitsandbytes==0.39.1
deepspeed==0.9.5
flash-attn==2.0.4
triton==2.0.0
einops==0.6.1
rotary-embedding-torch==0.2.3
